pyproject.toml
[project]
name = "bayhawk2-0"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.2",
    "uvicorn>=0.30.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic[email]>=2.0.0",
    "pydantic-ai>=1.75.0",
    "openai>=1.0.0",
    "ultralytics>=8.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "logfire>=4.31.0",
    "uvicorn>=0.42.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "aiosqlite>=0.20.0",
    "apscheduler>=3.10.4,<4",
    "supabase>=2.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.uv.scripts]
dev = { cmd = "uvicorn app.main:app --reload", env = { PYTHONPATH = "gateway" } }
test = { cmd = "pytest gateway/tests/ tests/test_pipeline.py -v", env = { IS_MOCK = "true", PYTHONPATH = "gateway", SCANNER_ENABLED = "false" } }

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
pythonpath = ["gateway"]
testpaths = ["gateway/tests"]
